requirements.txt
streamlit>=1.24.0
google-generativeai>=0.3.0
python-docx>=0.8.11
PyPDF2>=3.0.0
scikit-learn>=1.0.2
shap>=0.41.0
numpy>=1.21.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=0.19.0 